Package decomposed to distinct files (allow use on front without fs module)
